Change "download conversion compression level" type to Double
https://opensource.expediagroup.com/graphql-kotlin/docs/schema-generator/writing-schemas/scalars/#primitive-types

+++ server/src/main/kotlin/suwayomi/tachidesk/server/ServerConfig.kt
@@ -130,7 +130,7 @@ class ServerConfig(
 
     data class DownloadConversion(
         val target: String,
-        val compressionLevel: Float? = null,
+        val compressionLevel: Double? = null,
     )
 
     // extensions
